Changes in userincome UI, Added test for userincome

+++ userincome/views.py
@@ -23,15 +23,13 @@ def search_income(request):
 @login_required(login_url='/auth/login')
 def index(request):
     '''categories = Source.objects.all()'''
-    income = UserIncome.objects.filter(owner=request.user)
-    paginator = Paginator(income, 5)
+    income = UserIncome.objects.filter(owner=request.user).order_by('-id')
+    paginator = Paginator(income, 7)
     page_number = request.GET.get('page')
     page_obj = Paginator.get_page(paginator, page_number)
-    # currency = UserPreference.objects.get(user=request.user).currency
     context = {
         'income': income,
         'page_obj': page_obj,
-        # 'currency': currency
     }
     return render(request, 'income/index.html', context)
 
@@ -49,16 +47,9 @@ def add_income(request):
     if request.method == 'POST':
         amount = request.POST['amount']
 
-        if not amount:
-            messages.error(request, 'Amount is required')
-            return render(request, 'income/add_income.html', context)
         description = request.POST['description']
         date = request.POST['income_date']
         source = request.POST['source']
-
-        if not description:
-            messages.error(request, 'description is required')
-            return render(request, 'income/add_income.html', context)
 
         UserIncome.objects.create(owner=request.user, amount=amount, date=date,
                                   source=source, description=description)
@@ -77,20 +68,18 @@ def income_edit(request, id):
         'sources': sources
     }
     if request.method == 'GET':
-        return render(request, 'income/edit_income.html', context)
+        requested_user = request.user
+        if income.owner == requested_user:
+            return render(request, 'income/edit_income.html', context)
+        else:
+            return redirect('income')
     if request.method == 'POST':
         amount = request.POST['amount']
 
-        if not amount:
-            messages.error(request, 'Amount is required')
-            return render(request, 'income/edit_income.html', context)
         description = request.POST['description']
         date = request.POST['income_date']
         source = request.POST['source']
 
-        if not description:
-            messages.error(request, 'description is required')
-            return render(request, 'income/edit_income.html', context)
         income.amount = amount
         income. date = date
         income.source = source
@@ -104,6 +93,8 @@ def income_edit(request, id):
 
 def delete_income(request, id):
     income = UserIncome.objects.get(pk=id)
-    income.delete()
-    messages.success(request, 'record removed')
+    requested_user = request.user
+    if income.owner == requested_user:
+        income.delete()
+        messages.success(request, 'record removed')
     return redirect('income')
